+++ mex/backend/merged/helpers.py
@@ -3,9 +3,13 @@ from typing import Any, Literal, overload
 from pydantic import ValidationError
 
 from mex.backend.graph.connector import GraphConnector
-from mex.backend.graph.exceptions import InconsistentGraphError, NoResultFoundError
+from mex.backend.graph.exceptions import (
+    InconsistentGraphError,
+    NoResultFoundError,
+)
 from mex.backend.rules.transform import transform_raw_rules_to_rule_set_response
 from mex.common.exceptions import MergingError
+from mex.common.logging import logger
 from mex.common.merged.main import create_merged_item
 from mex.common.models import (
     EXTRACTED_MODEL_CLASSES_BY_NAME,
@@ -196,8 +200,18 @@ def get_merged_item_from_graph(identifier: Identifier) -> AnyMergedModel:
     )
     if result["total"] == 0:
         msg = "Merged item was not found."
-        raise NoResultFoundError(msg) from None
+        raise NoResultFoundError(msg)
     if result["total"] != 1:
         msg = "Found multiple merged items."
         raise InconsistentGraphError(msg)
     return merge_search_result_item(result["items"][0], Validation.STRICT)
+
+
+def delete_merged_item_from_graph(identifier: Identifier) -> None:
+    """Delete a merged item including all extracted items and rule-sets."""
+    connector = GraphConnector.get()
+    result = connector.delete_item(identifier)
+    if result["deleted_merged_count"] == 0:
+        msg = "Merged item was not found."
+        raise NoResultFoundError(msg)
+    logger.info("deleted item %s: %s", identifier, result.one())