+++ mex/backend/merged/main.py
@@ -5,13 +5,20 @@ from fastapi import APIRouter, Path, Query
 from fastapi.exceptions import HTTPException
 from starlette import status
 
-from mex.backend.graph.exceptions import NoResultFoundError
+from mex.backend.graph.connector import GraphConnector
+from mex.backend.graph.exceptions import DeletionFailedError, NoResultFoundError
 from mex.backend.merged.helpers import (
+    delete_merged_item_from_graph,
     get_merged_item_from_graph,
     search_merged_items_in_graph,
 )
+from mex.backend.rules.helpers import get_rule_set_from_graph
 from mex.backend.types import MergedType, ReferenceFieldName
-from mex.common.models import AnyMergedModel, PaginatedItemsContainer
+from mex.common.models import (
+    MERGED_MODEL_CLASSES_BY_NAME,
+    AnyMergedModel,
+    PaginatedItemsContainer,
+)
 from mex.common.types import Identifier, Validation
 
 router = APIRouter()
@@ -58,3 +65,28 @@ def get_merged_item(identifier: Annotated[Identifier, Path()]) -> AnyMergedModel
         return get_merged_item_from_graph(identifier)
     except NoResultFoundError as error:
         raise HTTPException(status.HTTP_404_NOT_FOUND) from error
+
+
+@router.delete(
+    "/merged-item/{identifier}", status_code=status.HTTP_204_NO_CONTENT, tags=["editor"]
+)
+def delete_merged_item(
+    identifier: Annotated[Identifier, Path()],
+    include_rule_set: Annotated[  # noqa: FBT002
+        bool,
+        Query(
+            description="Delete with rule-set or "
+            "fail if rule-set is present and this parameter is False."
+        ),
+    ] = False,
+) -> None:
+    """Delete one merged item for the given `identifier`."""
+    connector = GraphConnector.get()
+    if not connector.exists_item(identifier, list(MERGED_MODEL_CLASSES_BY_NAME)):
+        raise HTTPException(status.HTTP_404_NOT_FOUND)
+    if include_rule_set is False and get_rule_set_from_graph(identifier):
+        raise HTTPException(status.HTTP_412_PRECONDITION_FAILED)
+    try:
+        delete_merged_item_from_graph(identifier)
+    except DeletionFailedError as error:
+        raise HTTPException(status.HTTP_409_CONFLICT) from error